isotopomer file with only Isotopomer class
